+++ storage/base.py
@@ -9,7 +9,7 @@ from typing import Optional
 
 from astrbot.api import logger
 
-from ..errors import CloudUploadError, CurlNotFoundError
+from ..errors import CurlNotFoundError
 from .curl import find_curl